Add files via upload
Removed the ⊖ and ⊕ buttons at the bottom of the screen.
Add a title “Memorize!” to the top of the screen.
Add  3 theme buttons to UI, Make cards that appear when a theme button is touched in an unpredictable (i.e. random) order.
The theme-choosing buttons include an image representing the theme and text describing the theme stacked on top of each other vertically.